beforeQuery functional hook
